requirements.txt
streamlit
pandas
joblib
numpy
plotly
pillow
scikit-learn==1.3.2